+++ src/components/Avatar/Avatar.tsx
@@ -8,7 +8,7 @@ import { Loader } from 'decentraland-ui/dist/components/Loader/Loader'
 import { WearablePreview } from 'decentraland-ui/dist/components/WearablePreview/WearablePreview'
 import Edit from '../../assets/icons/Edit.svg'
 import { config } from '../../modules/config'
-import { View } from '../../types'
+import { View } from '../../utils/view'
 import { EDIT_PROFILE_URL } from './consts'
 import { Props } from './Avatar.types'
 import styles from './Avatar.module.css'
