+++ src/components/Pages/MainPage/MainPage.tsx
@@ -5,7 +5,7 @@ import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { Tabs } from 'decentraland-ui/dist/components/Tabs/Tabs'
 import { Loader } from 'decentraland-ui'
 import { locations } from '../../../modules/routing/locations'
-import { getView } from '../../../types'
+import { getView } from '../../../utils/view'
 import { Avatar } from '../../Avatar'
 import { PageLayout } from '../../PageLayout'
 import { ProfileInformation } from '../../ProfileInformation'
