+++ src/utils/view.ts
@@ -1,0 +1,13 @@
+export enum View {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  OWN = 'own',
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  OTHER = 'other'
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+}
+
+export const getView = (loggedInAddress?: string, currentAccountAddress?: string): View => {
+  return currentAccountAddress === loggedInAddress || (currentAccountAddress === undefined && loggedInAddress !== undefined)
+    ? View.OWN
+    : View.OTHER
+}
